+++ coupling/CouplingMDDefinitions.h
@@ -33,10 +33,10 @@
 #define TAG_FROM_MACRO2MD 500000
 
 #include "tarch/la/Vector.h"
+#include "tarch/utils/OstreamOperators.h"
 #include <map>
 #include <set>
 #include <vector>
-#include "tarch/utils/OstreamOperators.h"
 
 namespace coupling {
 /** This is an enum for the macroscopic solver ID
