+++ coupling/interface/impl/LAMMPS/LammpsMDSimulation.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "../impl/LAMMPS/USER-MAMICO/mamico_cell.h"
-#include "../impl/LAMMPS/USER-MAMICO/mamico_lammps_md_solver_interface.h"
-#include "MDSimulation.h"
+#include "USER-MAMICO/mamico_cell.h"
+#include "USER-MAMICO/mamico_lammps_md_solver_interface.h"
+#include "coupling/interface/MDSimulation.h"
 #include "lammps/input.h"
 #include "lammps/lammps.h"
 #include <mpi.h>
