+++ coupling/interface/impl/ls1/LS1MDSimulation.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "MDSimulation.h"
+#include "coupling/interface/MDSimulation.h"
 #include "coupling/interface/impl/ls1/LS1StaticCommData.h"
 #include "ls1/src/Simulation.h"
 #include "ls1/src/plugins/MamicoCoupling.h"
