+++ coupling/services/MultiMDCellService.h
@@ -85,8 +85,8 @@ public:
     // localNumberMDSimulations*topologyOffset/intNumberProcesses MD simulations
     // before the actual block of ranks
     for (unsigned int i = 0; i < _blockOffset; i++) {
-      _macroscopicCellServices[i] = createMacroscopicCellServiceMacroOnly(i, macroscopicSolverInterface, mdConfiguration, _multiMDService, mamicoConfiguration, 
-      (i / _localNumberMDSimulations) * _intNumberProcesses);
+      _macroscopicCellServices[i] = createMacroscopicCellServiceMacroOnly(i, macroscopicSolverInterface, mdConfiguration, _multiMDService, mamicoConfiguration,
+                                                                          (i / _localNumberMDSimulations) * _intNumberProcesses);
       if (_macroscopicCellServices[i] == NULL) {
         std::cout << "ERROR "
                      "coupling::services::MultiMDCellService::MultiMDCellServic"
@@ -115,8 +115,8 @@ public:
     // allocate all macroscopic cell services for macro-only-solver AFTER
     // topology offset
     for (unsigned int i = _blockOffset + _localNumberMDSimulations; i < _totalNumberMDSimulations; i++) {
-      _macroscopicCellServices[i] = createMacroscopicCellServiceMacroOnly(i, macroscopicSolverInterface, mdConfiguration, _multiMDService, mamicoConfiguration, 
-      (i / _localNumberMDSimulations) * _intNumberProcesses);
+      _macroscopicCellServices[i] = createMacroscopicCellServiceMacroOnly(i, macroscopicSolverInterface, mdConfiguration, _multiMDService, mamicoConfiguration,
+                                                                          (i / _localNumberMDSimulations) * _intNumberProcesses);
       if (_macroscopicCellServices[i] == NULL) {
         std::cout << "ERROR "
                      "coupling::services::MultiMDCellService::MultiMDCellServic"
@@ -702,18 +702,15 @@ private:
     return ic;
   }
 
-  coupling::services::MacroscopicCellService<dim>* createMacroscopicCellServiceMacroOnly(
-    unsigned int ID, 
-    coupling::interface::MacroscopicSolverInterface<dim>* macroscopicSolverInterface,
-    simplemd::configurations::MolecularDynamicsConfiguration& mdConfiguration,
-    tarch::utils::MultiMDService<dim>& multiMDService,
-    coupling::configurations::MaMiCoConfiguration<dim>& mamicoConfiguration,
-    unsigned int topologyOffset) const {
+  coupling::services::MacroscopicCellService<dim>*
+  createMacroscopicCellServiceMacroOnly(unsigned int ID, coupling::interface::MacroscopicSolverInterface<dim>* macroscopicSolverInterface,
+                                        simplemd::configurations::MolecularDynamicsConfiguration& mdConfiguration,
+                                        tarch::utils::MultiMDService<dim>& multiMDService,
+                                        coupling::configurations::MaMiCoConfiguration<dim>& mamicoConfiguration, unsigned int topologyOffset) const {
     return new coupling::services::MacroscopicCellServiceMacroOnly<dim>(
-      ID, macroscopicSolverInterface, mdConfiguration.getMPIConfiguration().getNumberOfProcesses(), multiMDService.getGlobalRank(),
-      mdConfiguration.getDomainConfiguration().getGlobalDomainSize(), mdConfiguration.getDomainConfiguration().getGlobalDomainOffset(),
-      mamicoConfiguration.getParallelTopologyConfiguration(), mamicoConfiguration.getMacroscopicCellConfiguration(),
-      topologyOffset);
+        ID, macroscopicSolverInterface, mdConfiguration.getMPIConfiguration().getNumberOfProcesses(), multiMDService.getGlobalRank(),
+        mdConfiguration.getDomainConfiguration().getGlobalDomainSize(), mdConfiguration.getDomainConfiguration().getGlobalDomainOffset(),
+        mamicoConfiguration.getParallelTopologyConfiguration(), mamicoConfiguration.getMacroscopicCellConfiguration(), topologyOffset);
   }
 
   unsigned int _localNumberMDSimulations;
